+++ broker/CHANGELOG.md
@@ -1,6 +1,15 @@
+# Unreleased
+
+## Breaking
+
+- The return type of `PendingConnection::client_data` is changed from `&SerializedValue` to
+  `&SerializedValueSlice`.
+
+
 # 0.2.0 (November 27th, 2023)
 
 - Bump for Aldrin 0.2.0.
+
 
 # 0.1.0 (November 24th, 2023)